Fix compiler warnings and document build warnings policy
- rtc_poller.c: move max_fd inside #ifndef _WIN32 (unused on Windows select)

- rtc_data_channel.c: cast RTC_DC_HEADER_SIZE to size_t in length check

- signaling_client.c: mark unused 'user' param in lws callback

Add Build Warnings Policy section to .github/copilot-instructions.md describing how to resolve common warning classes.

+++ rtc/src/rtc_data_channel.c
@@ -236,7 +236,7 @@ int rtc_dc_manager_recv(rtc_dc_manager_t *mgr, const uint8_t *data, size_t len) 
     /* data[3] is reserved */
     uint16_t payload_len = (uint16_t)((uint16_t)data[4] << 8 | data[5]);
 
-    if (RTC_DC_HEADER_SIZE + payload_len > len)
+    if ((size_t)RTC_DC_HEADER_SIZE + payload_len > len)
         return RTC_ERR_INVALID;
 
     const uint8_t *payload = data + RTC_DC_HEADER_SIZE;
